ci: move codecov publish from test script to circle pipeline (BROKEN)
This doesn't work, because the job runs in an entirely different container and the test coverage files don't exist there.

Leaving for learning. What I wrote before:

The script meant that npm test, if failing, would still exit with success because the codecov publish was the last command to be run.

It was also running on all branches in CI, rather than just master.